수정(시험): fit-test 자간 trim 시험을 tests/cases 로 옮긴다 (#5678)
source unit tier 정책이 제품 소스의 신규 `#[cfg(test)]` 를 금지한다. 앞 커밋이
`line_breaking.rs` 에 넣은 `fit_test_trim_tests` 가 CI Lint 의
`Validate Rust test suite manifest` 를 세웠다.

    PR base에 없는 신규 cfg(test) module 금지:
      src/renderer/composer/line_breaking.rs::fit_test_trim_tests#1
    source-side test 총량 증가 금지: 4226 > 4225

정책대로 다섯 건을 `tests/cases/issue_5678_fit_test_letter_spacing_trim.rs` 로
옮긴다. 주장은 그대로다.

헬퍼는 전부 private 이었으므로 `renderer::composer::fit_test_internals` 가
`#[doc(hidden)]` 으로 내보낸다 — 이 시험 하나 때문에 `pub` 이지 rhwp 의 API 가
아니다(`subsecond_dev` 선례).

`FitWidthHwp` 의 필드는 private 로 남긴다. "원시 정수는 생성자를 우회하지 못한다"는
#5678 의 계약이 밖에서도 그대로 서야 하므로, 값 읽기는 `hwp()`, 자간 보정 없는
대조군은 `untrimmed()` 로만 만든다.

음성 대조 재확인: trim 을 없애면 이전과 같은 2건
(`trimmed_width_follows_the_spacing_sign`,
`positive_spacing_trim_can_flip_the_fit_verdict`)이 실패한다.

검증: rust-test-suite-manifest --check, rust-unit-test-tiers --check (PR base 대비),
cargo fmt --check, cargo clippy -D warnings, clippy wasm32, 전체 cargo test 통과.
