+++ src/renderer/composer/line_breaking.rs
@@ -1159,8 +1159,12 @@ fn apply_paragraph_kerning_to_tokens(
 }
 
 /// px를 HWPUNIT(i32)로 변환 (내림, DPI=96 기준: px * 75)
+///
+/// `pub` 인 이유는 `tests/cases/issue_5678_fit_test_letter_spacing_trim.rs` 가 fit 판정 계약을
+/// 밖에서 구속하기 때문이지, rhwp 의 API 라는 뜻이 아니다. `#[doc(hidden)]` 이 그 사실을 적는다.
+#[doc(hidden)]
 #[inline]
-fn to_hwp(px: f64) -> i32 {
+pub fn to_hwp(px: f64) -> i32 {
     (px * 75.0) as i32
 }
 
@@ -1178,7 +1182,8 @@ fn condensed_line_width_hwp(width_hwp: i32, space_savings_hwp: i32) -> i32 {
 
 // 한컴은 HWPUNIT 정수 양자화 시 미세한 반올림 차이를 허용한다.
 // 15 HU 이내의 초과는 줄에 포함한다.
-const LINE_BREAK_TOLERANCE: i32 = 15;
+#[doc(hidden)]
+pub const LINE_BREAK_TOLERANCE: i32 = 15;
 
 fn condense_fit_can_pull_next_token(
     current_width_hwp: i32,
@@ -1215,7 +1220,8 @@ fn condense_fit_can_pull_next_token(
 /// `-0.16…-1.76` px).
 /// Forced to 0 under an active character grid, which is inert here: every
 /// corpus section has `char_grid == 0`.
-fn fit_test_letter_spacing_trim_hwp(letter_spacing_px: &[f64], token_end_idx: usize) -> i32 {
+#[doc(hidden)]
+pub fn fit_test_letter_spacing_trim_hwp(letter_spacing_px: &[f64], token_end_idx: usize) -> i32 {
     if token_end_idx == 0 {
         return 0;
     }
@@ -1258,26 +1264,43 @@ fn resolved_letter_spacing_px(
 /// 한 자리만 자간을 뺀 값을 넘겼다. 자간이 0 인 문단에서는 두 값이 같아 어떤 테스트도
 /// 차이를 잡지 못했다. 이제 원시 정수는 이 함수에 들어가지 못하고, 호출부는 생성자
 /// 이름으로 어느 쪽인지 밝혀야 한다.
+///
+/// `pub` 이지만 내부 필드는 private 이다 — 원시 정수가 생성자를 우회하지 못한다는 위 계약이
+/// 밖에서도 그대로 선다. `#[doc(hidden)]` 은 이것이 rhwp 의 API 가 아님을 적는다.
+#[doc(hidden)]
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
-struct FitWidthHwp(i32);
+pub struct FitWidthHwp(i32);
 
 impl FitWidthHwp {
+    /// fit 판정 폭을 HWPUNIT 으로 읽는다. 시험이 값을 확인하는 유일한 통로다.
+    #[inline]
+    pub fn hwp(self) -> i32 {
+        self.0
+    }
+
+    /// 자간 보정 없이 fit 판정 폭을 만든다 (자간이 0 인 문단·대조군용).
+    #[inline]
+    pub fn untrimmed(token_width_hwp: i32) -> Self {
+        Self(token_width_hwp)
+    }
+
     /// 후보 토큰의 마지막 글자 뒤 자간을 뺀 폭. 실사용 fill 이 쓰는 값이다.
     ///
     /// 줄 끝에 오는 글자의 뒤 자간은 그려지지 않으므로 들어가는지 따질 때 빼고 잰다.
     /// 펜은 전체 폭만큼 전진한다.
-    fn trimmed(token_width_hwp: i32, letter_spacing_px: &[f64], token_end_idx: usize) -> Self {
+    pub fn trimmed(token_width_hwp: i32, letter_spacing_px: &[f64], token_end_idx: usize) -> Self {
         Self(token_width_hwp - fit_test_letter_spacing_trim_hwp(letter_spacing_px, token_end_idx))
     }
 
     /// 커닝 경계쌍 보정을 fit 판정 폭에 더한다 (#4439 커닝 세션과의 병합점).
     /// 펜 전진 폭에는 더하지 않는다 — fit 판정 전용 축이다.
-    fn with_pair_adjustment(self, adjustment_hwp: i32) -> Self {
+    pub fn with_pair_adjustment(self, adjustment_hwp: i32) -> Self {
         Self(self.0 + adjustment_hwp)
     }
 }
 
-fn text_token_fits_line_hwp(
+#[doc(hidden)]
+pub fn text_token_fits_line_hwp(
     current_width_hwp: i32,
     token_width: FitWidthHwp,
     space_savings_hwp: i32,
@@ -4074,91 +4097,6 @@ fn compute_line_spacing_hwp(
             let min_hwp = round_px_to_hwpunit(ls_value, dpi);
             (min_hwp - line_height_hwp).max(0)
         }
-    }
-}
-
-#[cfg(test)]
-mod fit_test_trim_tests {
-    use super::*;
-
-    /// [#5678] 문제 3 — trim 이 어느 시험에도 구속되지 않고 글자마다 할당되고 있었다.
-    /// 부호 두 방향과 경계를 여기서 못박는다.
-    ///
-    /// 계약: **후보 토큰의 마지막 글자 뒤 자간만** 뺀다. 줄 끝 글자의 뒤 자간은 그려지지
-    /// 않으므로 "들어가는가"를 따질 때 빼고 재고, 펜은 전체 폭만큼 전진한다.
-    #[test]
-    fn trim_takes_only_the_spacing_after_the_candidate_last_char() {
-        let spacing = [1.0, 2.0, 4.0];
-        // token_end_idx 는 exclusive — 마지막 글자는 idx-1 이다.
-        assert_eq!(fit_test_letter_spacing_trim_hwp(&spacing, 1), to_hwp(1.0));
-        assert_eq!(fit_test_letter_spacing_trim_hwp(&spacing, 2), to_hwp(2.0));
-        assert_eq!(fit_test_letter_spacing_trim_hwp(&spacing, 3), to_hwp(4.0));
-    }
-
-    /// 토큰이 비었거나(`0`) 자간 배열 밖이면 보정하지 않는다.
-    #[test]
-    fn trim_is_zero_outside_the_spacing_slice() {
-        let spacing = [3.0];
-        assert_eq!(fit_test_letter_spacing_trim_hwp(&spacing, 0), 0);
-        assert_eq!(fit_test_letter_spacing_trim_hwp(&spacing, 9), 0);
-        assert_eq!(fit_test_letter_spacing_trim_hwp(&[], 1), 0);
-    }
-
-    /// **부호가 고정돼 있지 않다.** 양수 자간은 후보를 좁게, 음수 자간은 넓게 만든다.
-    ///
-    /// 이슈가 지적한 대로 근거 코퍼스(`76076_regulatory_analysis`)는 `-0.16…-1.76px`
-    /// 로 음수 방향뿐이었다. 양수 방향을 여기서 함께 고정한다.
-    #[test]
-    fn trimmed_width_follows_the_spacing_sign() {
-        let w = to_hwp(100.0);
-        let narrower = FitWidthHwp::trimmed(w, &[2.0], 1);
-        let wider = FitWidthHwp::trimmed(w, &[-2.0], 1);
-        assert!(
-            narrower.0 < w,
-            "양수 자간은 fit 판정 폭을 좁혀야 한다: {} vs {w}",
-            narrower.0
-        );
-        assert!(
-            wider.0 > w,
-            "음수 자간은 fit 판정 폭을 넓혀야 한다: {} vs {w}",
-            wider.0
-        );
-        assert_eq!(narrower.0, w - to_hwp(2.0));
-        assert_eq!(wider.0, w - to_hwp(-2.0));
-    }
-
-    /// **양수 자간에서 trim 이 판정을 뒤집는 지점이 실재한다.**
-    ///
-    /// 이슈 문제 2 가 지목한 상쇄다 — 자기 뒤 자간을 뺀 덕에만 들어간 토큰이 있고,
-    /// 펜은 `w_hwp` 전체만큼 전진한다. 이것은 결함이 아니라 **선언된 계약**이다
-    /// (줄 끝 자간은 그려지지 않는다). 다만 계약이 실제로 발동하는 구간이 있다는
-    /// 사실 자체를 고정해 두어, 나중에 trim 을 없애도 아무 시험이 안 깨지는 일이
-    /// 다시 생기지 않게 한다.
-    #[test]
-    fn positive_spacing_trim_can_flip_the_fit_verdict() {
-        let effective = to_hwp(100.0);
-        let current = to_hwp(90.0);
-        // 자연 폭은 tolerance 를 넘고, trim 을 빼면 들어간다.
-        let token_w = effective + LINE_BREAK_TOLERANCE - current + to_hwp(1.0);
-        let untrimmed = FitWidthHwp(token_w);
-        let trimmed = FitWidthHwp::trimmed(token_w, &[2.0], 1);
-        assert!(
-            !text_token_fits_line_hwp(current, untrimmed, 0, effective, 12.0),
-            "trim 없이는 들어가지 않아야 한다"
-        );
-        assert!(
-            text_token_fits_line_hwp(current, trimmed, 0, effective, 12.0),
-            "trim 을 빼면 들어가야 한다"
-        );
-    }
-
-    /// 커닝 보정은 fit 판정 폭에만 더한다 — 펜 전진 폭 축과 섞이지 않는다.
-    #[test]
-    fn pair_adjustment_only_moves_the_fit_width() {
-        let w = to_hwp(50.0);
-        let base = FitWidthHwp::trimmed(w, &[1.0], 1);
-        let adjusted = base.with_pair_adjustment(to_hwp(3.0));
-        assert_eq!(adjusted.0, base.0 + to_hwp(3.0));
     }
 }
